Move `PrawnRails::Document` patching to a module
Closes #7

Also moved `FieldRow` to its own file to have a nice one class/module
per file mapping.
